[PREGEL] Bugfix/fix staging build (#17575)
* Delete tests that will be redundant anyway

They fail on mac

* Circumvent indirect conversion that fails on windows

+++ arangod/Pregel/CommonFormats.h
@@ -151,7 +151,7 @@ struct ColorPropagationValue {
     VectorOfColors result;
     auto const size = colors.size();
     result.reserve(size);
-    for (size_t color = 0; color < size; ++color) {
+    for (uint16_t color = 0; color < size; ++color) {
       if (contains(color)) {
         result.push_back(color);
       }
